Add addon structure
* Adds an addon structure where each addon will register itself
  against an addon map.

  * Each addon can register optional callbacks that will be called
    if present before and after applying the addon to the cluster.

  * These callbacks can be used to renew certificates, or perform
    any other actions.

  * By refactoring the addon struct to an addon map, remove several
    ad-hoc switches regarding addon logic, that checked what version
    should be returned depending on the requested addon.

* Refactors the code so each addon file contains everything related
  to that addon.

* Addons are applied locally as opposed to the previous general
  approach of scp'ing the manifests and applying them remotely.

  * While on `bootstrap` we have SSH access to the machine, during
    an addon upgrade SSH access is not desired, we should use the
    cluster as another consumer; for this reason, render the addons
    locally and apply them on the local machine with `kubectl`.

* `skuba cluster init` now iterates on the list of addons for the
  current Kubernetes version, and creates the addons that are
  registered for that version.

* Adds a `skuba-config` ConfigMap with an entry `SkubaConfiguration`,
  that holds a struct that gets (un)marshaled. This struct only
  contains one field right now: `AddonsVersion`, that is the map
  containing the current addons version for all deployed addons on
  the cluster. This is what is used to determine whether an addon
  should be updated.

  * Without this method, it's not trivial to decide in what exact
    version an addon is: consider an addon with several resources,
    should all resources be annotated? What happens if some of them
    are out of date while others are up to date, what version we
    choose? For this reason, a `skuba-config` ConfigMap holds the
    information about every addon version and manifest version.

  * Addons are always applied, unless the version hardcoded in skuba
    is less than the one present in the cluster (this can happen if
    a user is using an outdated `skuba` in a cluster where addons were
    upgraded with a newer `skuba`).

  * The ConfigMap is created upon `bootstrap` with the current version
    of all addons matching the Kubernetes version used to boostrap the
    cluster.

  * Every time an addon is applied, the ConfigMap gets updated with the
    new version and manifest version.

* Since the `tooling` image is not strictly an addon, move it to the
  `ComponentsVersion` struct, so we can still use different `tooling`
  images depending on the current cluster version if needed.

* Add `PSP` as a first class citizen addon: use an arbitrary "1.0.0"
  version since no images will be pulled in any case for these
  manifests, so the image version is not relevant.

* Addons have two Apply modes: `AlwaysRender` and
  `SkipRenderIfConfigFilePresent`.

  * `AlwaysRender` will always get the addon manifest template and
    render the last version of it, this is the desired behavior
    always except for a `skuba node bootstrap` operation.

  * `SkipRenderIfConfigFilePresent` will render the addon manifest
    template only if the config file is not present in disk. This
    is useful because we allow the user to run `skuba cluster init`,
    tweak the configuration if they want, and run
    `skuba node bootstrap` afterwards. In this case, we want to apply
    the user modified settings, instead of forcefully rerender.

Note well:

* When an addon has a new version or manifest revision (manifest
  revisions should *always* be bumped), it will be reapplied on the
  cluster when running `skuba addon upgrade apply`. This means that
  any user customization will be lost when performing this action.

  * A way to tackle this problem would be to allow for user provided
    `kustomize` patches that live inside the cluster; when we upgrade
    our manifests (the base), we run the user `kustomize` patches on
    top of them, applying the patched version to the cluster. This
    would ensure that we would be able to evolve our base manifests
    while keeping the user changes on every upgrade. Also, it should
    be possible to detect manifest conflicts early, by running the
    patch operation beforehand in a sandbox.
