+++ go.mod
@@ -17,6 +17,7 @@ require (
 	k8s.io/cluster-bootstrap v0.0.0
 	k8s.io/klog v0.3.1
 	k8s.io/kubernetes v1.15.0
+	sigs.k8s.io/yaml v1.1.0
 )
 
 replace (
