+++ internal/app/skuba/cluster/init.go
@@ -26,11 +26,7 @@ import (
 	"k8s.io/klog"
 	"k8s.io/kubernetes/cmd/kubeadm/app/images"
 
-	cilium "github.com/SUSE/skuba/internal/pkg/skuba/cni"
-	"github.com/SUSE/skuba/internal/pkg/skuba/dex"
-	"github.com/SUSE/skuba/internal/pkg/skuba/gangway"
 	"github.com/SUSE/skuba/internal/pkg/skuba/kubernetes"
-	"github.com/SUSE/skuba/internal/pkg/skuba/kured"
 	"github.com/SUSE/skuba/pkg/skuba"
 	cluster "github.com/SUSE/skuba/pkg/skuba/actions/cluster/init"
 )
@@ -61,22 +57,15 @@ func NewInitCmd() *cobra.Command {
 			}
 
 			initConfig := cluster.InitConfiguration{
-				ClusterName:         args[0],
-				CloudProvider:       initOptions.CloudProvider,
-				ControlPlane:        initOptions.ControlPlane,
-				PauseImage:          images.GetGenericImage(skuba.ImageRepository, "pause", kubernetes.ComponentVersionForClusterVersion(kubernetes.Pause, kubernetesVersion)),
-				CiliumImage:         cilium.GetCiliumImage(),
-				CiliumInitImage:     cilium.GetCiliumInitImage(),
-				CiliumOperatorImage: cilium.GetCiliumOperatorImage(),
-				KuredImage:          kured.GetKuredImage(),
-				DexImage:            dex.GetDexImage(),
-				GangwayClientSecret: dex.GenerateClientSecret(),
-				GangwayImage:        gangway.GetGangwayImage(),
-				KubernetesVersion:   kubernetesVersion,
-				ImageRepository:     skuba.ImageRepository,
-				EtcdImageTag:        kubernetes.ComponentVersionForClusterVersion(kubernetes.Etcd, kubernetesVersion),
-				CoreDNSImageTag:     kubernetes.ComponentVersionForClusterVersion(kubernetes.CoreDNS, kubernetesVersion),
-				StrictCapDefaults:   initOptions.StrictCapDefaults,
+				ClusterName:       args[0],
+				CloudProvider:     initOptions.CloudProvider,
+				ControlPlane:      initOptions.ControlPlane,
+				PauseImage:        images.GetGenericImage(skuba.ImageRepository, "pause", kubernetes.ComponentVersionForClusterVersion(kubernetes.Pause, kubernetesVersion)),
+				KubernetesVersion: kubernetesVersion,
+				ImageRepository:   skuba.ImageRepository,
+				EtcdImageTag:      kubernetes.ComponentVersionForClusterVersion(kubernetes.Etcd, kubernetesVersion),
+				CoreDNSImageTag:   kubernetes.ComponentVersionForClusterVersion(kubernetes.CoreDNS, kubernetesVersion),
+				StrictCapDefaults: initOptions.StrictCapDefaults,
 			}
 
 			err := cluster.Init(initConfig)
