+++ internal/pkg/skuba/dex/dex.go
@@ -22,30 +22,24 @@ import (
 	"fmt"
 
 	"github.com/pkg/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/kubernetes/cmd/kubeadm/app/constants"
-	"k8s.io/kubernetes/cmd/kubeadm/app/images"
 	"k8s.io/kubernetes/cmd/kubeadm/app/util/pkiutil"
 
 	"github.com/SUSE/skuba/internal/pkg/skuba/kubernetes"
+	"github.com/SUSE/skuba/internal/pkg/skuba/node"
 	"github.com/SUSE/skuba/internal/pkg/skuba/util"
-	"github.com/SUSE/skuba/pkg/skuba"
-	node "github.com/SUSE/skuba/pkg/skuba/actions/node/bootstrap"
 )
 
 const (
-	imageName = "caasp-dex"
-
-	CertCommonName = "oidc-dex"
-	SecretCertName = "oidc-dex-cert"
+	certCommonName = "oidc-dex"
+	secretCertName = "oidc-dex-cert"
 )
 
 // CreateCert creates a signed certificate for dex
 // with kubernetes CA certificate and key
-func CreateCert(
-	client clientset.Interface,
-	pkiPath, kubeadmInitConfPath string,
-) error {
+func CreateCert(client clientset.Interface, pkiPath, kubeadmInitConfPath string) error {
 	// Load kubernetes CA
 	caCert, caKey, err := pkiutil.TryLoadCertAndKeyFromDisk(pkiPath, constants.CACertAndKeyBaseName)
 	if err != nil {
@@ -60,23 +54,17 @@ func CreateCert(
 
 	// Generate dex certificate
 	cert, key, err := util.NewServerCertAndKey(caCert, caKey,
-		CertCommonName, cfg.ClusterConfiguration.APIServer.CertSANs)
+		certCommonName, cfg.ClusterConfiguration.APIServer.CertSANs)
 	if err != nil {
 		return errors.Wrap(err, "could not genenerate dex server cert")
 	}
 
 	// Create or update secret resource
-	if err := util.CreateOrUpdateCertToSecret(client, caCert, cert, key, SecretCertName); err != nil {
+	if err := util.CreateOrUpdateCertToSecret(client, caCert, cert, key, secretCertName); err != nil {
 		return errors.Wrap(err, "unable to create/update cert to secret")
 	}
 
 	return nil
-}
-
-// GetDexImage returns dex image registry
-func GetDexImage() string {
-	return images.GetGenericImage(skuba.ImageRepository, imageName,
-		kubernetes.CurrentAddonVersion(kubernetes.Dex))
 }
 
 // GenerateClientSecret returns client secret which is used by
@@ -89,4 +77,14 @@ func GenerateClientSecret() string {
 	b := make([]byte, 12)
 	rand.Read(b)
 	return fmt.Sprintf("%x", b)
+}
+
+func DexCertExists(client clientset.Interface) (bool, error) {
+	_, err := client.CoreV1().Secrets(metav1.NamespaceSystem).Get(secretCertName, metav1.GetOptions{})
+	return kubernetes.DoesResourceExistWithError(err)
+}
+
+func RestartPods(client clientset.Interface) error {
+	listOptions := metav1.ListOptions{LabelSelector: fmt.Sprintf("app=%s", certCommonName)}
+	return client.CoreV1().Pods(metav1.NamespaceSystem).DeleteCollection(&metav1.DeleteOptions{}, listOptions)
 }
