+++ internal/pkg/skuba/kubernetes/kubelet.go
@@ -23,15 +23,16 @@ import (
 
 	batchv1 "k8s.io/api/batch/v1"
 	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/util/version"
 	"k8s.io/kubernetes/cmd/kubeadm/app/images"
 
 	"github.com/SUSE/skuba/pkg/skuba"
 )
 
-func DisarmKubelet(node *v1.Node) error {
+func DisarmKubelet(node *v1.Node, clusterVersion *version.Version) error {
 	return CreateAndWaitForJob(
 		disarmKubeletJobName(node),
-		disarmKubeletJobSpec(node),
+		disarmKubeletJobSpec(node, clusterVersion),
 	)
 }
 
@@ -39,7 +40,7 @@ func disarmKubeletJobName(node *v1.Node) string {
 	return fmt.Sprintf("caasp-kubelet-disarm-%s", node.ObjectMeta.Name)
 }
 
-func disarmKubeletJobSpec(node *v1.Node) batchv1.JobSpec {
+func disarmKubeletJobSpec(node *v1.Node, clusterVersion *version.Version) batchv1.JobSpec {
 	privilegedJob := true
 	return batchv1.JobSpec{
 		Template: v1.PodTemplateSpec{
@@ -47,7 +48,7 @@ func disarmKubeletJobSpec(node *v1.Node) batchv1.JobSpec {
 				Containers: []v1.Container{
 					{
 						Name:  disarmKubeletJobName(node),
-						Image: images.GetGenericImage(skuba.ImageRepository, "skuba-tooling", CurrentAddonVersion(Tooling)),
+						Image: images.GetGenericImage(skuba.ImageRepository, "skuba-tooling", ComponentVersionForClusterVersion(Tooling, clusterVersion)),
 						Command: []string{
 							"/bin/bash", "-c",
 							strings.Join(
