+++ internal/pkg/skuba/node/testdata/cluster.conf
