+++ internal/pkg/skuba/node/testdata/invalid.conf
@@ -1,0 +1,1 @@
+apiVersion: kubeadm.k8s.io/orz
