--- pkg/skuba/actions/node/join/testdata/invalid.conf
@@ -1,1 +1,0 @@
-apiVersion: kubeadm.k8s.io/orz
