--- pkg/skuba/actions/node/join/testdata/join.conf
@@ -1,17 +1,0 @@
-apiVersion: kubeadm.k8s.io/v1beta1
-caCertPath: /etc/kubernetes/pki/ca.crt
-discovery:
-  bootstrapToken:
-    apiServerEndpoint: 10.84.152.198:6443
-    token: j2sxf2.s3q02ll7nvbkjjr0
-    unsafeSkipCAVerification: true
-  timeout: 5m0s
-  tlsBootstrapToken: ""
-kind: JoinConfiguration
-nodeRegistration:
-  criSocket: /var/run/crio/crio.sock
-  kubeletExtraArgs:
-    cni-bin-dir: /usr/lib/cni
-    hostname-override: worker-0
-    pod-infra-container-image: registry.suse.de/devel/caasp/4.0/containers/containers/caasp/v4/pause:3.1
-  name: worker-0
