+++ pkg/skuba/actions/node/remove/remove.go
@@ -53,6 +53,11 @@ func Remove(client clientset.Interface, target string, drainTimeout time.Duratio
 		}
 	}
 
+	currentClusterVersion, err := kubeadm.GetCurrentClusterVersion()
+	if err != nil {
+		return errors.Wrap(err, "could not retrieve the current cluster version")
+	}
+
 	targetName := node.ObjectMeta.Name
 
 	var isControlPlane bool
@@ -69,7 +74,7 @@ func Remove(client clientset.Interface, target string, drainTimeout time.Duratio
 		etcd.RemoveMember(node)
 	}
 
-	if err := kubernetes.DisarmKubelet(node); err != nil {
+	if err := kubernetes.DisarmKubelet(node, currentClusterVersion); err != nil {
 		fmt.Printf("[remove-node] failed disarming kubelet: %v; node could be down, continuing with node removal...\n", err)
 	}
 
@@ -78,12 +83,8 @@ func Remove(client clientset.Interface, target string, drainTimeout time.Duratio
 			return errors.Wrapf(err, "[remove-node] could not remove the APIEndpoint for %s from the kubeadm-config configmap", targetName)
 		}
 
-		if err := cni.CreateOrUpdateCiliumConfigMap(); err != nil {
+		if err := cni.CreateOrUpdateCiliumConfigMap(client); err != nil {
 			return errors.Wrap(err, "[remove-node] could not update cilium-config configmap")
-		}
-
-		if err := cni.AnnotateCiliumDaemonsetWithCurrentTimestamp(); err != nil {
-			fmt.Printf("[remove-node] could not annonate cilium daemonset: %v\n", err)
 		}
 	}
